Add files via upload
This script will do an nmap ip/web address scan of open ports, services and operating system of your target. The scan will print a text file of the OS and Services to the specified location(my location was /home/nmap)and then start an instance of Metasploit. The text files created are set to overwrite each time you run the script.